requirements.txt

pandas
textblob
matplotlib
seaborn
fpdf2
